# Performance improvements for benchmarks