This pull request contains improvements to the performance of the example crates in this repository.

## fannkuch_redux

- I noticed a small performance reduction because LLVM was emitting memory accesses to the `self.odd` field, insted of storing it in a register.
  Not sure why this happens, LLVM should know we borrow `self` mutably & have exclusive access to `self` and be able to cache its value locally.

- Overall, about 1.5% better performance.

## stencil

See issue #95 for the full discussion.

- Now we're using `mul_adde` (FMA) where possible.
- Improved performance of `Data::reinit` by using unsafe memory accesses. Not part of the benchmark, but it was slow and showing up a lot in the profiler.
- The parallel code is slower than the SIMD code now:
```
simd: 175 ms
simd+par: 279 ms
ispc: 176 ms
ispc+tasks: 373 ms
```
  I guess that's... good?

### Failing tests
- I've updated the testing code to use approximative float comparisons, with an error no larger than 10^-4.

## aobench

- Added `mul_adde` to most places, matching ISPC's assembly.
- Inlined some functions which weren't getting inlined before.
- Overall, when using the system math library (instead of comparing against ISPC's built-in one) Rust performance is better than with ISPC.