@gnzlbg 2018-09-03
I think this can be updated to use the new `fmae` function right?