@GabrielMajeri 2018-09-03
@gnzlbg I was going to wait for #123 to land so I can use `mul_adde` .

I'm not planning to get this merged however, not until we manage to beat ISPC in `aobench` at least.