@GabrielMajeri 2018-09-04
I'd decided to compare oranges to oranges and use `libm` for both ISPC and Rust, to avoid comparing using `libsleef` vs. ISPC's inlined math functions.

The results are in:
```
cargo run --release 1920 1080 --algo tiled
time: 4170 ms
```
```
cargo run --release 1920 1080 --algo ispc
time: 4532 ms
```
So Rust is faster than ISPC. I don't know if it's due to a lack of inlining, or some issue with sleef, but it's clear there's not much more we can do to optimize the Rust code itself.