@gnzlbg 2018-09-04
Note that ISPC inline math functions are not as precise as sleef, libm, SVML, etc. 

Let me know when you finish the nitpicks and this is green so that we can merge this. I think it looks really good. I'll run the benchmarks on some systems and update the readme's afterwards (you can also add the results for your own systems in a follow up PR if you want).