@GabrielMajeri 2018-09-04
Alright, just fixed a _little_ **big** issue with the RNG: those functions were _not_ marked as `#[inline(always)]`, so the `target_feature` attribute would end up not applying to them, so we would end up with AVX2 code calling into a function built with SSE2 instructions.

<hr/>

I think I've addressed most of the issues I could, some final bechmark results here (with `libm`, in order from slowest to fastest):
```
ispc: 4542 ms
vector: 4355 ms
tiled: 4117 ms

ispc tasks: 1227 ms
vector_par:  1233 ms
tiled_par: 1170 ms
```

<hr/>

Side note: would it be fair to modify `build.rs` to force ISPC to always use the system's math library? After all, this repository is not about benchmarking's ISPC fast math library