@GabrielMajeri 2018-09-04
@gnzlbg Clippy failed, but I don't think it's related to this PR.

> I'd prefer if you do this in a different PR, because it applies to many benchmarks and I'd like to merge this ;)

Alright, I'll see if I can't get cross-language inlining to work, maybe we can use sleef to beat ISPC's math library.