@gnzlbg 2018-09-04
> Clippy failed, but I don't think it's related to this PR.

I'll fix that :) 

> Alright, I'll see if I can't get cross-language inlining to work, maybe we can use sleef to beat ISPC's math library.

I have a branch where I started doing this the other day. I can upload it if you want to pick it up from there.