@gnzlbg 2018-09-04
@GabrielMajeri i'm gnzlbg on IRC